Docker should be up and running. Refactored code.
