+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "backend"
+version = "0.1.0"
+description = ""
+authors = ["Viktor Kertanov <thelonius19@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.9"
+Flask = "^2.0.2"
+arrow = "^1.2.1"
+Flask-SQLAlchemy = "^2.5.1"
+typer = "^0.4.0"
+psycopg2-binary = "^2.9.2"
+
+[tool.poetry.dev-dependencies]
+mypy = "^0.930"
+wemake-python-styleguide = "^0.16.0"
+Faker = "^10.0.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
