--- requirements.txt
@@ -1,13 +1,0 @@
-arrow==1.2.1
-Flask==2.0.2
-Flask_SQLAlchemy==2.4.3
-fuzzywuzzy==0.18.0
-pandas==1.3.4
-typer==0.4.0
-psycopg2-binary==2.9.2
-
-# dev dependencies
-mypy==0.921
-wemake-python-styleguide==0.16.0
-
-Faker==10.0.0
