+++ setup.cfg
@@ -26,6 +26,7 @@ ignore =
     WPS410,  # disable: __all__ at init py not allowed
     WPS412,  # false-positive: init py with logic
     WPS428,  # ...
+    WPS430,  # false-positive: route + nested function in flask app
     WPS431,  # nested classes
     WPS462,  # ...
     WPS531,  # false-positive: simplified if return
@@ -55,4 +56,7 @@ check_untyped_defs = True
 disallow_untyped_defs = False
 
 [mypy-flask_sqlalchemy.*]
+ignore_missing_imports = True
+
+[mypy-sqlalchemy.*]
 ignore_missing_imports = True
