+++ weedly/config.py
@@ -1,5 +1,6 @@
 import os
 
-DOCKER_DB_URI = os.environ['DB_URL']
-SQLALCHEMY_DATABASE_URI = os.environ['DB_URL']
+DOCKER_DB_URI = os.environ['DB_URI']
+SQLALCHEMY_DATABASE_URI = os.environ['DB_URI']
+# SQLALCHEMY_DATABASE_URI = "postgresql://weedly:weedly-pass@db:5432/weedly"
 SQLALCHEMY_TRACK_MODIFICATIONS = False
