+++ weedly/db/db.py
@@ -1,6 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, scoped_session, declarative_base
 from weedly.config import DOCKER_DB_URI
+# from weedly.config import SQLALCHEMY_DATABASE_URI
 
 engine = create_engine(DOCKER_DB_URI)
 db_session = scoped_session(sessionmaker(bind=engine))
