+++ weedly/db/repos/postgres.py
@@ -1,43 +1,11 @@
-import random
-from abc import ABC, abstractmethod
 from sqlalchemy.sql import func
 from sqlalchemy import desc, inspect
 from sqlalchemy.exc import NoResultFound
 from typing import Dict, Any
 import json
 
-
 from weedly.db.db import db_session
-from weedly.db.model import News
-
-from weedly.tools.fake.create import get_fake_row_dict
-
-
-class Storage(ABC):
-    @property
-    @abstractmethod
-    def get_last_id(self):
-        raise NotImplementedError
-
-    @abstractmethod
-    def add(self, payload):
-        raise NotImplementedError
-
-    @abstractmethod
-    def delete(self, index):
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_one(self, index):
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_all(self, num_rows):
-        raise NotImplementedError
-
-    @abstractmethod
-    def update(self, index, payload):
-        raise NotImplementedError
+from weedly.db.repos.storage import Storage
 
 
 class PostgreStorage(Storage):
@@ -140,67 +108,3 @@ class PostgreStorage(Storage):
             for q in query_result:
                 final_list.append(self._row_as_dict(q))
             return json.dumps(final_list)
-
-if __name__ == "__main__":
-    "WARNING! SHITTY TESTS BELOW"
-    storage = PostgreStorage(News)
-    def p_line():
-        print("-"*50)
-    p_line()
-    num_rows = 10
-    print(f"We'll print out first {num_rows}")
-    for i, n in enumerate(storage.get_all(num_rows)):
-        print(f"{i+1}. {n}")
-
-    p_line()
-    print("We'll get the last index")
-    max_index = storage.get_last_id
-    print(max_index)
-    p_line()
-
-    last_index_news = storage.get_latest_by_id()
-    for i in last_index_news:
-        print(i)
-
-    id_to_get = 20
-    print(f"News have headers: {storage.table_headers}")
-
-    print(f"ID {id_to_get} contains the news: {storage.get_one(id_to_get)}")
-
-    urls = ["http://www.ooo.com/main/",
-            "https://rao.ru/terms/",
-            "http://ip.com/",
-            "https://meduza.io"
-            ]
-    for u in urls:
-        print(storage._url_in_table(u))
-
-    for _ in range(3):
-        fake_row = get_fake_row_dict()
-        print(f"Now we'll try to add a test article in the db: {fake_row}")
-        storage.add(fake_row)
-
-    for _ in range(3):
-        print("TESTING EXISTING URLS")
-        fake_row = get_fake_row_dict()
-        fake_row["url"] = random.choice(["http://ao.com/login/", "http://www.ooo.com/main/", "https://rao.ru/terms/", "http://ip.com/"])
-        # fake_row["author"] = None
-        storage.add(fake_row)
-
-    "Deleting a row test"
-    storage.delete(8)
-
-    """TESTING _id_in_table FUNCTION"""
-
-    print(f"Result = {storage._id_in_table(100)}")
-
-    print("-"*150)
-    print("We'll now test update method!")
-    fake_row = get_fake_row_dict()
-    index = 9
-    print(f"Our fake row to insert in index {index}: {fake_row}")
-    row_to_update = storage.get_one(9)
-
-    storage.update(index, fake_row)
-
-    print('Hello world!')
